reffs-main/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.0
celery==5.3.4
redis==5.0.1
Pillow==10.1.0
python-dotenv==1.0.0
gunicorn==21.2.0
whitenoise==6.6.0
django-filter==23.3
django-storages==1.14.2
boto3==1.28.64
psycopg2-binary==2.9.9
dj-database-url==2.1.0 